+++ tests/word_tokenizer/test_sentencepiece_tokenizer.py
@@ -4,22 +4,18 @@ from konoha.konoha_token import Token
 from konoha.word_tokenizer import WordTokenizer
 
 
-def test_kytea_with_s3_model():
+def test_word_tokenize_with_sentencepiece():
     try:
-        import boto3
-        del boto3
+        import sentencepiece
+        del sentencepiece
     except ImportError:
-        pytest.skip("skip s3 test because of missing boto3")
+        pytest.skip("Sentencepiece is not installed.")
 
-    try:
-        tokenizer = WordTokenizer(
-            tokenizer="KyTea",
-            model_path="s3://konoha-demo/kytea/model.knm"
-        )
-    except ImportError:
-        pytest.skip("skip kytea")
-
-    expect = [Token(surface=w) for w in "吾輩は 猫である".split(" ")]  # NOQA
+    tokenizer = WordTokenizer(
+        tokenizer="Sentencepiece",
+        model_path="data/model.spm"
+    )
+    expect = [Token(surface=w) for w in "▁ 吾 輩 は 猫 である".split(" ")]
     result = tokenizer.tokenize("吾輩は猫である")
     assert expect == result
 
@@ -37,7 +33,7 @@ def test_sentencepiece_with_s3_model():
             model_path="s3://konoha-demo/sentencepiece/model.spm"
         )
     except ImportError:
-        pytest.skip("skip sentencepiece")
+        pytest.skip("Sentencepiece is not installed.")
 
     expect = [Token(surface=w) for w in "▁ 吾 輩 は 猫 である".split(" ")]  # NOQA
     result = tokenizer.tokenize("吾輩は猫である")
